+++ src/Pim/Bundle/CatalogBundle/Form/Handler/AssociationTypeHandler.php
@@ -1,0 +1,79 @@
+<?php
+
+namespace Pim\Bundle\CatalogBundle\Form\Handler;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Form\FormInterface;
+use Doctrine\Common\Persistence\ObjectManager;
+use Pim\Bundle\CatalogBundle\Entity\AssociationType;
+
+/**
+ * Form handler for association type
+ *
+ * @author    Filips Alpe <filips@akeneo.com>
+ * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class AssociationTypeHandler
+{
+    /**
+     * @var FormInterface
+     */
+    protected $form;
+
+    /**
+     * @var Request
+     */
+    protected $request;
+
+    /**
+     * @var ObjectManager
+     */
+    protected $manager;
+
+    /**
+     * Constructor for handler
+     * @param FormInterface $form    Form called
+     * @param Request       $request Web request
+     * @param ObjectManager $manager Storage manager
+     */
+    public function __construct(FormInterface $form, Request $request, ObjectManager $manager)
+    {
+        $this->form    = $form;
+        $this->request = $request;
+        $this->manager = $manager;
+    }
+
+    /**
+     * Process method for handler
+     * @param AssociationType $associationType
+     *
+     * @return boolean
+     */
+    public function process(AssociationType $associationType)
+    {
+        $this->form->setData($associationType);
+
+        if ($this->request->isMethod('POST')) {
+            $this->form->submit($this->request);
+
+            if ($this->form->isValid()) {
+                $this->onSuccess($associationType);
+
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Call when form is valid
+     * @param AssociationType $associationType
+     */
+    protected function onSuccess(AssociationType $associationType)
+    {
+        $this->manager->persist($associationType);
+        $this->manager->flush();
+    }
+}