+++ src/Pim/Bundle/CatalogBundle/Form/Type/ProductAssociationType.php
@@ -22,10 +22,10 @@ class ProductAssociationType extends AbstractType
     {
         $builder
             ->add(
-                'association',
+                'associationType',
                 'oro_entity_identifier',
                 array(
-                    'class'    => 'Pim\Bundle\CatalogBundle\Entity\Association',
+                    'class'    => 'Pim\Bundle\CatalogBundle\Entity\AssociationType',
                     'property' => 'id',
                     'multiple' => false
                 )